REV-001: Update the program entry point to handle arguments safely and config location.
